# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistem Koordinasi Bencana Bali API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relawan:
    get:
      operationId: getRelawan
      tags: [relawan]
      summary: Daftar relawan
      parameters:
        - name: status_verifikasi
          in: query
          schema:
            type: string
            enum: [menunggu, terverifikasi, ditolak]
      responses:
        "200":
          description: Daftar relawan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Relawan"
    post:
      operationId: createRelawan
      tags: [relawan]
      summary: Daftar sebagai relawan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelawanInput"
      responses:
        "201":
          description: Pendaftaran relawan berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relawan"
  /relawan/{id}:
    get:
      operationId: getRelawanById
      tags: [relawan]
      summary: Detail relawan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detail relawan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relawan"
  /relawan/{id}/verifikasi:
    patch:
      operationId: verifikasiRelawan
      tags: [relawan]
      summary: Verifikasi relawan oleh admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifikasiRelawanInput"
      responses:
        "200":
          description: Status verifikasi diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relawan"
...
components:
  schemas:
...
    Relawan:
      type: object
      required: [id, nama, nik, no_hp, alamat, keahlian, status_verifikasi, created_at]
      properties:
        id:
          type: integer
        nama:
          type: string
        nik:
          type: string
        no_hp:
          type: string
        email:
          type: ["string", "null"]
        alamat:
          type: string
        keahlian:
          type: array
          items:
            type: string
        foto_ktp_url:
          type: ["string", "null"]
        status_verifikasi:
          type: string
          enum: [menunggu, terverifikasi, ditolak]
        catatan_verifikasi:
          type: ["string", "null"]
        created_at:
          type: string
    RelawanInput:
      type: object
      required: [nama, nik, no_hp, alamat, keahlian]
      properties:
        nama:
          type: string
        nik:
          type: string
        no_hp:
          type: string
        email:
          type: ["string", "null"]
        alamat:
          type: string
        keahlian:
          type: array
          items:
            type: string
        foto_ktp_url:
          type: ["string", "null"]
    VerifikasiRelawanInput:
      type: object
      required: [status_verifikasi]
      properties:
        status_verifikasi:
          type: string
          enum: [terverifikasi, ditolak]
        catatan_verifikasi:
          type: ["string", "null"]
...