# lib/api-zod/src/generated/types/relawan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Koordinasi Bencana Bali API
 * OpenAPI spec version: 0.1.0
 */
import type { RelawanStatusVerifikasi } from './relawanStatusVerifikasi';

export interface Relawan {
  id: number;
  nama: string;
  nik: string;
  no_hp: string;
  /** @nullable */
  email?: string | null;
  alamat: string;
  keahlian: string[];
  /** @nullable */
  foto_ktp_url?: string | null;
  status_verifikasi: RelawanStatusVerifikasi;
  /** @nullable */
  catatan_verifikasi?: string | null;
  created_at: string;
}